Add Loading splashscreen
