+++ js/src/dapps/index.ejs
@@ -5,9 +5,35 @@
     <meta name="viewport" content="width=device-width">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <title><%= htmlWebpackPlugin.options.title %></title>
+    <style>
+      html, body, #container {
+        width: 100%;
+        height: 100%;
+        margin: 0;
+        padding: 0;
+      }
+
+      .loading-container {
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        height: 100%;
+        font-family: Roboto;
+        background-color: rgba(0, 0, 0, 0.8);
+        color: #ddd;
+      }
+
+      .loading {
+        font-size: 4em;
+      }
+    </style>
   </head>
   <body>
-    <div id="container"></div>
+    <div id="container">
+      <div class="loading-container">
+        <span class="loading">Loading...</span>
+      </div>
+    </div>
     <script src="vendor.js"></script>
     <% if (!htmlWebpackPlugin.options.secure) { %>
       <script src="/parity-utils/parity.js"></script>
